% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2.R
\name{process_GBD_data}
\alias{process_GBD_data}
\title{GBD数据读取与处理主函数}
\usage{
process_GBD_data(
  ZIP_folder = FALSE,
  file_fold = TRUE,
  Parallel = FALSE,
  foldername
)
}
\arguments{
\item{ZIP_folder}{是否读取zip压缩包（TRUE/FALSE）}

\item{file_fold}{是否直接读取csv文件夹（TRUE/FALSE）}

\item{Parallel}{是否启用并行读取csv}

\item{foldername}{zip或csv所在的文件夹路径}
}
\value{
返回一个标准化后的GBD数据框
}
\description{
一键读取、标准化、合并GBD数据
}
